* Minor adjustment to DPAP components

+++ jolivia.protocol/src/main/java/org/dyndns/jkiddo/dmap/chunks/picture/ImageFileSize.java
@@ -33,7 +33,7 @@ public class ImageFileSize extends UIntChunk
 	 * @param filesize
 	 *            The filesize of the image.
 	 */
-	public ImageFileSize(int filesize)
+	public ImageFileSize(long filesize)
 	{
 		super("pifs", "dpap.imagefilesize", filesize);
 	}
